Refactor command
